[Darwin] Retrieve data stored into ~/Documents/chip.store (#14233)

+++ scripts/tests/chiptest/test_definition.py
@@ -19,6 +19,8 @@ import time
 from datetime import datetime
 import typing
 import threading
+from pathlib import Path
+import platform
 
 from enum import Enum, auto
 from dataclasses import dataclass
@@ -99,12 +101,26 @@ class TestDefinition:
                     "Unknown test target - don't know which application to run")
 
             tool_cmd = paths.chip_tool
-            if os.path.exists('/tmp/chip_tool_config.ini'):
-                os.unlink('/tmp/chip_tool_config.ini')
+
+            files_to_unlink = [
+                '/tmp/chip_tool_config.ini',
+                '/tmp/chip_tool_config.alpha.ini',
+                '/tmp/chip_tool_config.beta.ini',
+                '/tmp/chip_tool_config.gamma.ini',
+            ]
+
+            for f in files_to_unlink:
+                if os.path.exists(f):
+                    os.unlink(f)
 
             # Remove server all_clusters_app or tv_app storage, so it will be commissionable again
-            if os.path.exists('/tmp/chip_kvs'):
-                os.unlink('/tmp/chip_kvs')
+            if platform.system() == 'Linux':
+                if os.path.exists('/tmp/chip_kvs'):
+                    os.unlink('/tmp/chip_kvs')
+
+            if platform.system() == "Darwin":
+                if os.path.exists(str(Path.home()) + '/Documents/chip.store'):
+                    os.unlink(str(Path.home()) + '/Documents/chip.store')
 
             discriminator = str(randrange(1, 4096))
             logging.debug(
@@ -139,3 +155,4 @@ class TestDefinition:
         finally:
             if app_process:
                 app_process.kill()
+                app_process.wait(3)
